% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{CodonCountPos}
\alias{CodonCountPos}
\title{Create a file with Codon count numbers (adjust by position) and a file with Codon usage}
\usage{
CodonCountPos(filenm, CodonPerc, xlsxFS, f, metds = 1)
}
\arguments{
\item{infile}{"DNA sequence file", "Codon Usage Table file", "Result Excel file and "Input file serial number"}
}
\value{
None (Result files with "*_codonSEQ.txt" and "_codonUSTB" and Excel sheets will be created)
}
\description{
Create a file with Codon count numbers (adjust by position) and a file with Codon usage
}
