% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{MotifCount}
\alias{MotifCount}
\title{Create a file with Motif and Restrict Sites sequence count numbers}
\usage{
MotifCount(filenm, motifnm)
}
\arguments{
\item{infile}{DNA sequence file and Motif define file}
}
\value{
None (new file with "*_motifCount.txt" will be created)
}
\description{
Create a file with Motif and Restrict Sites sequence count numbers
}
