% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{RareCodonCalc}
\alias{RareCodonCalc}
\title{Create a file with Rare codon and GC contents information}
\usage{
RareCodonCalc(filenm, CodonPercN)
}
\arguments{
\item{infile}{DNA sequence file and New Codon Usage Table file}
}
\value{
None (new file with "*_RareCD.txt" will be created)
}
\description{
Create a file with Rare codon and GC contents information
}
