% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{RareCodonGraph}
\alias{RareCodonGraph}
\title{Create one graph file for each DNA sequence related with rare Codon information}
\usage{
RareCodonGraph(filenm, CodonPercN, f)
}
\arguments{
\item{infile}{DNA sequence file and New Codon Usage Table file}
}
\value{
None (One or more PDF file(s) with related DNA sequence ID will be created)
}
\description{
Create one graph file for each DNA sequence related with rare Codon information
}
