% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{TransProtein}
\alias{TransProtein}
\title{Create a file with protein sequence in the third column}
\usage{
TransProtein(filenm, CodonPerc, xlsxFS, f)
}
\arguments{
\item{infile}{"DNA sequence file", "Codon Usage Table file", "Result Excel file and "Input file serial number"}
}
\value{
None (Result file with "*_AddAAS.txt" and Excel sheet "AAseq" will be created)
}
\description{
Create a file with protein sequence in the third column
}
