% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{MotifCount}
\alias{MotifCount}
\title{Create a file with Motif and Restrict Sites sequence count numbers}
\usage{
MotifCount(filenm, motifnm, xlsxFS, f)
}
\arguments{
\item{infile}{"DNA sequence file", "Motif define file", "Result Excel file and "Input file serial number"}
}
\value{
None (Result file with "*_motifCount.txt" and Excel sheet will be created)
}
\description{
Create a file with Motif and Restrict Sites sequence count numbers
}
