% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{normCodonPerc}
\alias{normCodonPerc}
\title{Create a new Codon usage table file with normalized Codon usage values on last colum}
\usage{
normCodonPerc(CodonPerc)
}
\arguments{
\item{infile}{Codon Usage Table file}
}
\value{
None (new file with "*N.txt" will be created)
}
\description{
Create a new Codon usage table file with normalized Codon usage values on last colum
}
