% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{RareCodonCalc}
\alias{RareCodonCalc}
\title{Create a file with Rare codon and GC contents information}
\usage{
RareCodonCalc(filenm, CodonPercN, xlsxFS, f)
}
\arguments{
\item{infile}{"DNA sequence file", "New codon Usage Table file", "Result Excel file and "Input file serial number"}
}
\value{
None (Result file with "*_RareCD.txt"  and Excel sheet will be created)
}
\description{
Create a file with Rare codon and GC contents information
}
