% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDS_ANA.R
\name{TransProtein}
\alias{TransProtein}
\title{Create a file with protein sequence in the third column}
\usage{
TransProtein(filenm, CodonPerc)
}
\arguments{
\item{infile}{DNA sequence file and Codon Usage Table file}
}
\value{
None (new file with "*_AddAAS.txt" will be created)
}
\description{
Create a file with protein sequence in the third column
}
